Remove _config.yml from cache manifest
Github pages by default does not serve files starting with underscore.
Including one in the cache manifest leads to a 404 condition. A very
confusing situation as it is plainly there in the repository.

Another caveat is that at least Chrome won't actually say _which_ file it is
that cannot be downloaded. It will actually display the 404 for the
app.manifest file itself.